pyproject.toml
[project]
name = "emrakul"
version = "0.1.0"
description = "Agent orchestration framework - Claude as orchestrator, specialized workers for execution"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "redis>=5.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
emrakul = "emrakul.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]  # Line length handled separately for SSH commands

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]
